Improved arithmetic in CoreWarmupMonitor::getClockSpeedRatio().
By representing the minClockTick_ as a double-precision number of nanoseconds,
the arithmetic for computing the normalized clock speed is simplified.

Also, added a tolerance when checking for clock speed drops, since it seems
there's sometimes a little instability in the clock speed.  Perhaps this
will need further tweaking, or even to be exposed as a parameter.