Improved overhead modeling of Benchmark::istream_string* tests.
Since these operations involve ovewriting the string's existing value,
it seems that should be added to the overhead subtracted off.
The reason to regard it as overhead is that its main cost is deallocation,
which is primarily a function of the string's old contents rather than
what's being parsed.  Furthermore, in typical usage, the target of the
operation will tend to be empty.