Added Category::string_to benchmarks.
Note that there are two sets of string_to benchmarks, for the floats and
doubles.  Here, the idea was to have a set of numbers directly comparable
to the corresponding istream tests, which meant the input needed to be
formatted the same way (i.e. via std::ostringstream).  However, I also
wanted a set of tests indicating the cost of parsing these numbers the
way that std::to_string() formats them. Hence, the _os variants
correspond to the former, while the _ts correspond to the latter.