Added option: --coreB=N - Which core to use for secondary thread.
Also, if secondary thread's core is unset or explicitly -1, then a specific core
is now chosen before calling SetCoreAffinity() to lock down the primary thread.
This turns out to be necessary for enabling auto-selection of the secondary
thread - otherwise, the secondary thread inherits the affinity of the primary
and always lands on the same core (unless the user explicitly specified a
different one).

Since external factors (such as taskset) can affect availability of cores,
autoselection of the secondary thread's core has been limited to 3 attempts.
The user will be warned, if this limit has been exceeded.