Enhanced struct ScopedFile & modified directory_iter_ + readdir_ to use subdir.
Changed ScopedFile::ScopedFile() to create an empty, unbound instance.
This is mostly useful for overloaded constructors and static factory functions.

Added static member functions ScopedFile::make_random() and make_random_in(),
to make it more explicit, when a file with a unique name is being created,
as well as providing control over where.

Enhanced ScopedFile with the ability to work with directories, if the
O_DIRECTORY is specified.  This takes it beyond the functionality of a simple
open()/close()/unlink() wrapper, but I think this behavior is still quite
reasonable and should be generally unsurprising.

Changed Benchmark::directory_iter_4k and Benchmark::readdir_* to create a
private subdirectory, so they're not affected by any files in the current one.
These changes motivated most of those, above.